Update Django and dependencies
Also upgrade Django project structure by comparing it with a new Django 2.2 LTS generated project. Also set the timezone to NZ.

Since this is a fairly small project, not much needed to be changed for Django 2.2.

The django-phonenumber-field dependency now expects the user to add the dependency phonenumbers yourself, so added this to requirements.txt